wip: added test to entities, review consumer teste

+++ api/tests/BillingTest/Service/MailServiceTest.php
@@ -88,7 +88,7 @@ class MailServiceTest extends TestCase
             'email' => 'charlesdickens@kanastra.com.br',
             'amount' => '10.00',
             'due_date' => '2023-03-30',
-            'debt_id' => (int)6566,
+            'debt_id' => 6566,
         ];
     }
 }
